BookFollowing + Light adjustments
Book is no longer a child subject of the player , so we could have an easier time mantaining book's position and just adding player's position manually through the inspector. We can adjust how fast and far the book travels. I also adjusted the light, so now there is a permanent sun.